more packages! (more info below)
adding lots of random packages to package.json.

we defo don't need all of these but the app is barely hanging on by a thread. if some packages will help it keep its will to live, who are we to dispossess it of the opportunity.

don't know what stuff package-lock.json is high on

~ the expo ones are for directing app. the rest are an attempt to circumvent a bug. for local testing you may remove them (other than safe-area-context for it is used in app.js)